refactor(sanctions-pep): address CodeRabbitAI review comments
- Extract Zod schemas from test files to contracts.ts production module
- Extract business logic functions from test files to business-logic.ts production module
- Update index.ts to import from production modules instead of test files
- Wrap all handler logic in try/catch blocks with descriptive error messages
- Add environment variable validation before server startup
- Fix GET curl examples in README.md to use query parameters instead of JSON body
- Fix integration test type issues: replace 'any' with concrete types
- Update JurisdictionRiskInputSchema to require at least one jurisdiction
- Update integration tests to assert x-payment-required header in 402 responses
